moved wait helper (#14)
